[CLI] Keep skill generation in skills.py; pass content into _skills
Undo the module move that ballooned the PR diff: build_skill_md() and
its helpers stay in skills.py exactly as on main. skills.py passes the
generated content into _skills.py as a value instead:
- install_generated_skill(content, dest) writes SKILL.md + marker
- update_skills() takes a required keyword-only hf_cli_content
This keeps every import at module top with no skills/_skills cycle and
the PR diff strictly on-topic.
